Add Time class
This class was created to encapsulate fractional time values into the
same object as regular time values, and to simplify methods that perform
time math, especially those that involve fractional time calculations.

+++ src/main/java/net/lavabucket/hourglass/time/effects/RandomTickSleepEffect.java
@@ -51,7 +51,7 @@ public class RandomTickSleepEffect extends AbstractTimeEffect {
         int speed = SERVER_CONFIG.baseRandomTickSpeed.get();
         SleepStatus sleepStatus = context.getTimeService().sleepStatus;
         if (condition == ALWAYS || (condition == SLEEPING && !sleepStatus.allAwake())) {
-            speed *= context.getTimeDelta();
+            speed *= context.getTimeDelta().longValue();
         }
 
         context.getTimeService().level.setRandomTickSpeed(speed);
